(chore) http client healthcheck tests
